Catch+recover errors from HTTP request failures in worker handshake
When, in performing a worker handshake, the HTTP request (using the `request` module) reports an error, we now simply report an unsuccessful handshake, rather than propagating an onError notification. This allows the `pickWorker()` function to continue on to the next IP to check, rather than ending all attempts to pick a worker.