Client.js: Update console.info call to print viz worker URL
In client.js:requestWorker(), we print the address of the worker we were routed to using console.info(). This commit updates that call to use the vizaddr API call's new return value (a URL object, not a String containing an address).

We also remove the printing of the time elapsed for that call to complete. For some reason, the reply.data.timestamp property was not being included in vizaddr's return value, even though the server should definitely be sending it. My guess is that 'timestamp' is some reserved name somewhere (Express' return value, the JSON serializer/deserializer, XHR, etc.)