Update to use unified log level option in config & not use process.env
Previously, the config module had two options, one for the log level to the console, and one for the log level to the log file. logger.js also checked process.env for overrides to these values.

This commit updates logger.js to use the new config API, which only has a single log level option (which it sets based off of config.ENVIRONMENT). logger also no longer checks process.env for overrides to these values, since that should be the purview of the config module.

Additionally, we make the logic of 'log to STDOUT if config.BUNYAN_LOG is undefined, otherwise log to to the file config.BUNYAN_LOG' clearer here.