Enable Bunyan built-in serializers
We use a custom serializer for the `err` field of logged objects, and overwrite the built-in Bunyan serializer in bunyan.stdSerializers.err with it.

Previously, we would then pass `{err: bunyan.stdSerializers.err}` as the value of the `serializers` option to `bunyan.createLogger()`. This caused Bunyan to only have one serializer (for `err`, using `bunyan.stdSerializers.err`).

We now pass all of `bunyan.stdSerializers` as the value of the `serializers` option, which causes Bunyan to use all of its built-in serializers in addition to our custom err serializer.