Make StreamGL a devDependency, not a standard dependency
StreamGL is used iff graph-viz is run as a standalone program, rather than loaded by viz-server. In fact, the `require('StreamGL')` statement is guarded by a `if (require.main === module)`. In other words, we only use it when running/debugging graph-viz locally.

By making StreamGL a devDependency, we break a circular dependency that StreamGL now has on viz-server (for running integration tests.) Now, StreamGL will still be installed if you do a `npm install` in the graph-viz directory (as when you install and run it locally,) but won't be installed when graph-viz is installed as a dependency of StreamGL.