Forbid SlackBot from accessing any of our URLs
Slack's link preview generator bot may be causing problems with central and/or nginx. Specifically, it may be causing HTTP 416 errors ("Requested Range Not Satisfiable") by sending bad HTTP range headers. This *may* be the cause of the bug where loading graph.html hangs loading the static resources of the page. We see Express printing an "Error: Requested Range Not Satisfiable" message + stack trace to STDERR at the least. This may indicate an unhandled exception that is causing us havoc, or may just be Express being verbose with its logging.

Either way, we don't need (or really want) Slack generating previews for our app URLs. This commit adds a directive to the nginx config files which checks if the HTTP user-agent header starts with "Slack". If so, it returns a HTTP 403 response ('Error: forbidden').

According to Slack's documentation (https://api.slack.com/robots), Slackbot specifically does not respect `robots.txt` or other methods of telling it to take a hike, so filtering at the HTTP level seems to be our only option.