Renderer disables vertexAttribArrays (Desc Below).
This seems to fix a performance bug that occured on the draw calls following a render of arrows. It sounds like disableVertexAttribArray isn't supposed to be strictly needed for WebGL (which is true, it works correctly), but it appears to have performance impacts.This could be due to OSX's underlying OpenGL implementation.