Move var workers_port_start from roles/worker to group_vars/app_servers
By making `workers_port_start` a 'global' var (well, a group var for the group `app_servers`, but since all our servers are app servers...) we can easily access it from multiple roles, rather than just inside the `worker` role. Specifically, this facilities upcoming changes to the `proxy` role which make use of this variable.

This commit also sets a default value for the variable in inside the `worker` and `proxy` roles (in the `defaults` variable file for the respective roles). This will make sure the two roles that use `workers_port_start` won't blow up if we forget to set that variable elsewhere.