central: don't install nginx reverse proxy as part of the central role
Currently, we install nginx reverse proxies in two places in our deploy scripts: in the `central` role, and in the `proxy` role (the former for when you're accessing the instance directly, where nginx proxies to localhost, and the latter when you're accessing the <server>-secure.graphistry.com servers, where nginx will proxy to another instance).

This commit moves us toward only installing a reverse proxy in a single place, in the `proxy` role. Future commits will apply this role to `central` instances, and do away with distinct <server>-secure.graphistry.com servers.