proxy: detect remote host's DNS server & store for use in nginx config
We need to configure the DNS server in nginx. Previously, we assumed we were deploying to ec2, and used its DNS server. This commit changes it so we grab the DNS server configured in the remote host's `/etc/resolv.conf`, and store that to an Ansible variable. If that fails for some reason, we fall back to using Google's public DNS server (8.8.8.8).

This commit does not currently make use of the DNS server variable. That will come in a later commit.