Apply `proxy` role to central; remove proxy server from inventory
Previously, we had a separate ec2 instance acting as our reverse proxy server (at least for staging), and our Ansible playbooks applied to `proxy` role to this server to deploy our nginx reverse proxy config to this server.

This commit changes it so that the "cluster"'s central server now also acts as the proxy server, doing away with a separate instance entirely. The playbooks now apply the `proxy` role to central. The proxy server is also removed from the inventory, and the plays in the playbook pertaining to that server specifically are also removed.