users: when creating users, don't bother to explicitly set their shell
Previously, we explicitly passed the `shell=/bin/bash` argument to Ansible's `user` module, to set the user's shell when we created the user. I don't really know why we did this, and it didn't seem to benefit us, so this commit stops doing it. (It also didn't seem to be causing problems that I know if, but less code now means fewer problems later).