worker: in reaper.py, comment out Boundary logging code
Since Boundary got bought by IBM, they've made a lot of breaking changes to how to log to it. We haven't kept on top of these, and don't even really use Boundary's dashboards anymore.

To quiet down the warnings we got from reaper.py trying to log to the old, outdated Boundary endpoints, we simply comment out Boundary logging in reaper.py. The previously logged info included GPU memory usage and number of workers active/idle. If we later decide Boundary is useful and want to revive logging, we can uncomment and update this code.