Proxy: allow apps to see when request was made with HTTPS (fixes #56)
Previously, our backend Node.js server apps would always have the protocol of all incoming requests set to `http`, even if the request made by the user was made using `https`. This made it impossible for app code to detect when a request was secure (for some value of 'secure'). This commit allows for Express.js to properly detect the connection protocol, and set `req.protocol` appropriately.

The reason for the bug is that our backend apps only support HTTP. It's our nginx reverse proxy that sits in front of all our apps that is handles HTTPS. When an HTTPS connection comes in, nginx unwraps it, and proxies it to the backend as plain HTTP. This is generally accepted Industry Best Practices™, since nginx deals with all the complicated security stuff, and there is less for our app code to screw up.

Because all requests to our backends are made from nginx, as HTTP, Express.js would always detect, and set, the `protocol` of the connection to `http`, even if the user connected to our site (i.e., our nginx reverse proxy) using HTTPS. This made it impossible for our app code to do things like reject sensitive actions if they were made over HTTP rather than HTTPS.

To fix this, we have the nginx reverse proxy set the `X-Forwarded-Proto` header of the HTTP request it makes to our backends, to be the protocol of the original request the user made to nginx. Express.js has built-in support for this header and (so long as `trust proxy` is true in Express' options) will set `req.protocol` to its value, rather than using the protocol nginx uses when talking to the backend (which is HTTP).