site.yml: change `remote_servers` group to `all`
We originally created the `remote_servers` host set so that we could apply roles to all servers, and exclude `localhost`. However, I've discovered that using the built-in host set `all` accomplishes the same thing, even ensuring `localhost` is excluded. So remove the `remote_servers` set, and replace it with `all`, in order to clean things up a bit.