reaper.py: parse argv for program options (db URL, hostname)
The python `argparse` option is used to define a set of command-line arguments to `reaper.py`. It then parses `argv` to find the supplied arguments, to resolve a final set of options for reaper.py to use at runtime.

At present, though, these options are not used; they exist in parallel with the existing methods of defining program options. A later commit with replace the existing options with this new method.

The goal of this is to remove all used of Ansible template variables, so we can deploy reaper.py without Ansible. Since we are setting program options using template vars, this adds an alternate way to set/get those options.