splunkfwd: copy .deb file from this repo, instead of downloading it
There was a bug that caused downloading the Splunk Forwarder from splunk.com to fail. Splunk's website was apparently recently updated, and their new SSL/TLS settings are incompatible with the version of Python installed on Ubuntu 14.04. Since the `get_url` Ansible module uses Python's `urllib2` internally, this caused the "Download Splunk Universal Forwarder" task to fail.

This commit applies the easy fix, and adds the Splunk forwarder .deb file to this repo. It then replaces the `get_url` task with a `copy` task, so that the Splunk forwarder is copied from the local system instead of downloaded from the website. The subsequent installation tasks proceed without change.