Replace Graphistry logo in corner of graph viz with SVG version of same
In the graph viz interface, we have the Graphistry logo and (optionally) the version number displayed in the lower-right corner.

This commit replaces the .png image we use for the Graphistry logo with a .svg version. The image should look exactly the same, but by using .svg instead of .png, we ensure our logo always looks crisp and professional, no matter what screen resolution or image size we are using.