In requisitionWorker, disable clientId checking for now
This is a mechanism for ensuring that only the user assigned to that worker by central is allowed to connect to it. However, central doesn't currently have an implementation of the requisite code, which meant client connections to the worker wouldn't have a clientId attached to them. This caused every connection to be rejected.

For the moment, we disable this checking in viz-app, meaning every client connection is accepted. Once we update central, we'll re-enable the checks, so that we increase overall security.