Disable "remove request body" functionality for the moment
The `proxy_set_header` directive in these blocks was causing nginx to ignore all the other `proxy_set_header` directives in the `server` block. This meant that all the proxy headers we were setting were not getting passed on.

For the moment, I've commented out the "remove request body" functionality. This fixes the header problem. Once we get stuff working on deploy, I'll circle back to this.