Nginx: fix bug causing WebSockets to be closed prematurely
I don't know why we *have* to send 'Connection: close' to the upstream from nginx for WebSockets to work, but my experiments show that if we don't, then the WebSocket connection from the browser is closed prematurely. This causes visualizations with large datasets to hang, because the WS connection disconnects while the dataset loads.