Use `X-Graphistry-Prefix` header for constructing worker URLs
To accommodate nginx routing requests to a particular worker, we need to prefix absolute URLs with `/worker/1235`, and set the `<base>` tag to add that prefix to the 'actual' page URL (so that relative URLs will use that prefixed path). This logic exists in `src/viz-worker/middleware/redner.js`.

This commit changes the header we use for getting that prefix and base URL. The headers are set by us in our nginx config, as a way of letting the worker know its prefix. We updated our nginx config to tweak those headers recently, and this commit modifies viz-app to account for those tweaks.