Fix bug causing etl requests to fail
Nginx has undocumented behavior where, when a request is redirected internally via a `X-Accel-Redirect` header, the request has its method forcibly set to "GET". Central routes worker requests (like `/etl` and `/graph/...`) to the requisitioned worker using that header to tell nginx to redirect the request.

`/etl` requests are *supposed* to be POST requests (since the client is sending data to the server). However, when the request was redirected, nginx was changing them to GET requests. This caused the viz-app server to error out when it saw an ETL request without any data (the request body having been discarded when the request method was changed).

After a ton of research and digging through the nginx source code, I discovered an additional undocumented behavior: if the location being redirected to by the `X-Accel-Redirect` header is a "named" location block (a location block starting with "@"), then the request method isn't changed. However, named location blocks have a lot of caveats, such as being excluded from normal request handling; we have to redirect to such a block using an nginx directive that supports such a thing.

This commit changes all of our worker location blocks to be named. For example, "/worker/1234" was changed to "@worker1234". Now, central just set `X-Accel-Redirect` to "@worker1234", and nginx will redirect the request to that worker, without messing with it.

Now, we still need to route incoming requests to this worker. We can't exactly send a request for a URI like "GET @worker1234/graph/graph.html", so we need to redirect legal URIs, like "/worker/1234/graph/graph.html" to the "@worker1234" location block.

Our new config grabs the worker ID ("1234" in the above example) from the request path, and maps that to a worker location block, like "@worker1234" using a `map` directive. We then have a `try_files` directive that tries to find & respond with a file we know doesn't exist (`/workers_offline.html`), and pass the named worker location as the final, "fallback" location for `try_files`. This will always be triggered, since we know the file never exists. Thus, we do an internal redirect to a named location.