Add script to auto-generate `routes-workers.conf` for some # of workers
We don't need to run this script on deploy or anything, it's just a convenience when we want to quickly change the # of workers and/or their config.

Note that `upstreams-workers.conf` isn't covered by this script, so you still need to manually define each worker upstream and it's IP:port, inside that file.