Back out config changes to viz-app’s ‘root’ `index.js` file
(The below is per a conversation with @trxcllnt)

Paul’s `dev/header-toolbar-split-pane` branch changes the main server include architecture, in order to have a better method of hot reloading. This new method negates the need to pass the `node-convict`-based config to submodules from the root index. Therefore, this commit backs out the changes I previously made on this branch to do that (which were done to work around some weird hot reloading logic).