In logger, tweak how HTTP requests are logged to remove sensitive data
Bug DEV-1237 describes how, when our apps have authorization enabled, we are inadvertently writing the authorized username and password to our logs on every HTTP request.

The cause of this is that we’re logging every HTTP request that comes in using Bunyan’s built-in `req` object serializer. This is great for analyzing app/user behavior and debugging problems, and is similar to how NGINX and other web servers also log every request that they process.

However, Bunyan’s built-in `req` object serializer will log all the headers sent in the request, along with the rest of the request’s properties. The headers are mostly benign (things like the request’s user agent, cache-control headers, etc.), with the major exception of the `authorization` header. Because our apps currently use HTTP Basic Authentication, this header contains the username and password of the user, in plaintext (well, base64 encoded, but not hashed/encrypted). This header, with the username and password, is sent with every HTTP request to the server, meaning we will write the username and password to our logs on every HTTP request.

This commit fixes the bug so that we omit the username+password from our logs. Specifically, we override the `req` object serializer in the Bunyan logger instance(s) we create to use our own implementation, rather than Bunyan’s standard built-in implementation. Our `req` serializer simply calls the built-in `req` serializer, then takes the serialized object it returns, removes a pre-set list of headers from its header list, and returns this redacted object for Bunyan to use as the log entry.

To start with here, we remove the `authorization` header that contains the request’s username+password. This fixes the immediate bug, and means we no longer write the username+password to the logs. In the future, we can add more headers to the list of headers we remove from the logs, if we decide that those headers contain sensitive information (for example, the cookie header).