Upgrade @graphistry/common dep to v3.1.0 to fix leaking password in log
This commit upgrades our `@graphistry/common` dependency from v3.0.2 to v3.1.0. `common` is the (private Graphistry) npm module that contains our logger, among other functionality.

v3.1.0 of that module includes a fix for bug DEV-1237, which was causing us to write the HTTP authentication credentials of incoming requests (which includes the app password) to the log file. The new version of `common` now omits that header when logging incoming HTTP requests, which should fix the problem.