In the list of investigations, make the links act like links
This commit modifies the investigation table on the main screen. In that table, the name of each investigation is wrapped in a `<a>` link element, and clicking it takes you to that investigation.

Previously, the `<a>` element had no `href=“...”` attribute. Instead, we added a `onclick` handler to the `<a>` element that would change the screen to the investigation screen for that particular investigation. While this worked, it broke link semantics: a user couldn’t right-click the link and open it in a new tab/window, or copy the link address for saving/sharing, or see what URL would be pulled up by mousing over the link.

This commit adds a `href=“...”’ attribute to the `<a>` element, with the static URL of the investigation. We still use the `onclick` to do the actual navigation; this allows us to change the page without telling the browser to navigate to a new URL. The `href` is there strictly for appearances; the user can now open the link in a new tab, copy it, see the URL when hovering, etc.

The `onclick` handler has been modified to call `e.preventDefault()` to cancel the normal ‘navigate to the page specified in the `href` attribute’ behavior in the browser, since we do our own thing.