Add skeleton code to allow us to upload this to npm as a module
This adds `package.json` to allow us to treat this repo as an npm module, including uploading it to npm as "@graphistry/docs-update", and installing it as a dependency of other modules. This allows us to serve this from `central`.

I also added `index.js`, which just exports a variable (`docsUpdateAssetPath`) which gives the absolute path on the filesystem of the module, for easier serving via express's `static` middleware by modules that import this module.