fix rendering-scheduler render loop race conditions
1. Make `startRenderingLoop` function schedule its first execution of `loop` on the animationFrame, instead of immediately executing it
2. delete tasks from the renderQueue as they're rendered, instead of wiping out the renderQueue Object after each loop
3. store the deferred "slow" render tasks in a special slow queue, to be processed at the very end once the render loop "settles" into a quiet state

There's complex interdependency between the render tasks that doesn't need to exist, and I believe is still contributing to some performance degradation in edge rendering. This commit doesn't resolve that, but it does fix the race conditions that were causing edges not to update when the render loop settles (which was the same root cause as the edges not rendering when color encodings changed).